Move datadogRoute and thirdPartyRoute under request folder
