+++ packages/base/src/commands/elf-symbols/helpers.ts
@@ -2,7 +2,7 @@ import type {RequestBuilder} from '@datadog/datadog-ci-base/helpers/interfaces'
 import type {MultipartPayload, UploadOptions} from '@datadog/datadog-ci-base/helpers/upload'
 
 import {getBaseSourcemapIntakeUrl} from '@datadog/datadog-ci-base/helpers/base-intake-url'
-import {datadogRoute} from '@datadog/datadog-ci-base/helpers/datadog-route'
+import {datadogRoute} from '@datadog/datadog-ci-base/helpers/request/datadog-route'
 import {upload} from '@datadog/datadog-ci-base/helpers/upload'
 import {getRequestBuilder} from '@datadog/datadog-ci-base/helpers/utils'
