+++ packages/base/src/commands/git-metadata/library.ts
@@ -2,9 +2,9 @@ import type {CommitInfo} from './interfaces'
 import type {SimpleGit} from 'simple-git'
 
 import {newApiKeyValidator} from '../../helpers/apikey'
-import {datadogRoute} from '../../helpers/datadog-route'
 import type {RequestBuilder} from '../../helpers/interfaces'
 import {Logger, LogLevel} from '../../helpers/logger'
+import {datadogRoute} from '../../helpers/request/datadog-route'
 import type {UploadOptions} from '../../helpers/upload'
 import {upload, UploadStatus} from '../../helpers/upload'
 import {getRequestBuilder, filterAndFormatGithubRemote} from '../../helpers/utils'
