+++ packages/base/src/commands/git-metadata/upload.ts
@@ -6,7 +6,6 @@ import {Command, Option} from 'clipanion'
 import {FIPS_ENV_VAR, FIPS_IGNORE_ERROR_ENV_VAR} from '../../constants'
 import type {ApiKeyValidator} from '../../helpers/apikey'
 import {newApiKeyValidator} from '../../helpers/apikey'
-import {datadogRoute} from '../../helpers/datadog-route'
 import {toBoolean} from '../../helpers/env'
 import {InvalidConfigurationError} from '../../helpers/errors'
 import {enableFips} from '../../helpers/fips'
@@ -15,6 +14,7 @@ import type {RequestBuilder} from '../../helpers/interfaces'
 import {Logger, LogLevel} from '../../helpers/logger'
 import type {MetricsLogger} from '../../helpers/metrics'
 import {getMetricsLogger} from '../../helpers/metrics'
+import {datadogRoute} from '../../helpers/request/datadog-route'
 import {UploadStatus} from '../../helpers/upload'
 import {getRequestBuilder, timedExecAsync} from '../../helpers/utils'
 import {cliVersion} from '../../version'
