+++ packages/base/src/commands/measure/measure.ts
@@ -1,5 +1,4 @@
 import type {CILevel} from '@datadog/datadog-ci-base/helpers/ci'
-import type {RequestError} from '@datadog/datadog-ci-base/helpers/request'
 
 import chalk from 'chalk'
 import {Command, Option} from 'clipanion'
@@ -8,12 +7,14 @@ import {BaseCommand} from '@datadog/datadog-ci-base'
 import {FIPS_ENV_VAR, FIPS_IGNORE_ERROR_ENV_VAR} from '@datadog/datadog-ci-base/constants'
 import {getDatadogSite} from '@datadog/datadog-ci-base/helpers/api'
 import {LEVEL_TO_NUMBER, enrichCIEnvFromGithubLogs, getCIEnv, validateLevel} from '@datadog/datadog-ci-base/helpers/ci'
-import {datadogRoute} from '@datadog/datadog-ci-base/helpers/datadog-route'
 import {toBoolean} from '@datadog/datadog-ci-base/helpers/env'
 import {enableFips} from '@datadog/datadog-ci-base/helpers/fips'
 import {retryRequest} from '@datadog/datadog-ci-base/helpers/retry'
 import {parseMeasuresFile} from '@datadog/datadog-ci-base/helpers/tags'
 import {getApiHostForSite, getRequestBuilder} from '@datadog/datadog-ci-base/helpers/utils'
+
+import type {RequestError} from '../../helpers/request'
+import {datadogRoute} from '../../helpers/request/datadog-route'
 
 export const parseMeasures = (measures: string[]) =>
   measures.reduce((acc, keyValue) => {
