+++ packages/base/src/commands/tag/tag.ts
@@ -5,10 +5,10 @@ import {FIPS_ENV_VAR, FIPS_IGNORE_ERROR_ENV_VAR} from '../../constants'
 import {getDatadogSite} from '../../helpers/api'
 import type {CILevel} from '../../helpers/ci'
 import {LEVEL_TO_NUMBER, enrichCIEnvFromGithubLogs, getCIEnv, validateLevel} from '../../helpers/ci'
-import {datadogRoute} from '../../helpers/datadog-route'
 import {toBoolean} from '../../helpers/env'
 import {enableFips} from '../../helpers/fips'
 import type {RequestError} from '../../helpers/request'
+import {datadogRoute} from '../../helpers/request/datadog-route'
 import {retryRequest} from '../../helpers/retry'
 import {parseTags, parseTagsFile} from '../../helpers/tags'
 import {getApiHostForSite, getRequestBuilder} from '../../helpers/utils'
