+++ packages/base/src/commands/trace/api.ts
@@ -1,7 +1,7 @@
 import type {Payload} from './interfaces'
 import type {RequestBuilder} from '@datadog/datadog-ci-base/helpers/interfaces'
 
-import {datadogRoute} from '@datadog/datadog-ci-base/helpers/datadog-route'
+import {datadogRoute} from '@datadog/datadog-ci-base/helpers/request/datadog-route'
 import {getRequestBuilder} from '@datadog/datadog-ci-base/helpers/utils'
 
 export const reportCustomSpan = (request: RequestBuilder) => async (customSpan: Payload) => {
