+++ packages/base/src/helpers/request/third-party-route.ts
@@ -1,0 +1,5 @@
+declare const thirdPartyRouteBrand: unique symbol
+
+export type ThirdPartyRoute = string & {[thirdPartyRouteBrand]: true}
+
+export const thirdPartyRoute = (url: string): ThirdPartyRoute => url as ThirdPartyRoute
